refactor: apiBuilder에서 맵퍼만 남기고 제거 후 파일 이름 변경
